backend/requirements.txt
Django==2.2.19
django-colorfield==0.9.0
mypy==1.4.1
mypy-extensions==1.0.0
python-dotenv==1.0.0
# pytz==2023.3
# sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
asgiref==3.7.2
django-filter==23.2
djangorestframework==3.14.0
# djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Pillow==10.0.0
drf-extra-fields==3.5.0 
# filetype-1.2.0
